+++ app/Http/Controllers/Auth/LogoutController.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
+
+class LogoutController extends Controller
+{
+    // TODO Turn these functions into stateless-friendly endpoints ie. meaningful status codes and json data
+
+    public function __invoke(): RedirectResponse
+    {
+        // Logout current device and clear session
+        Auth::logout();
+
+        $request = request();
+
+        // Invalidate user's session and regenerate CSRF token
+        $request->session()->invalidate();
+        $request->session()->regenerateToken();
+
+        return redirect()->route('home');
+    }
+}