+++ resources/views/dashboard.blade.php
@@ -1,1 +1,6 @@
 Hey, {{ auth()->user()->username }}
+
+<form action="{{ route('logout') }}" method="POST">
+    @csrf
+    <button type="submit">Logout</button>
+</form>