+++ resources/views/home.blade.php
@@ -1,0 +1,1 @@
+Welcome to {{ config('app.name') }}