+++ routes/auth.php
@@ -2,12 +2,18 @@
 
 use App\Http\Controllers\Auth\EmailVerificationController;
 use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\Auth\LogoutController;
 use App\Http\Controllers\Auth\RegisterController;
 use Illuminate\Support\Facades\Route;
 
-// Email verification routes
-Route::prefix('auth')->group(function () {
-    Route::prefix('verify-email')->middleware('auth')->group(function () {
+// Logout and Email verification routes
+Route::prefix('auth')->middleware('auth')->group(function () {
+
+    // Logout route
+    Route::post('logout', LogoutController::class)->name('logout');
+
+    // Email verification routes
+    Route::prefix('verify-email')->group(function () {
 
         Route::get('', [EmailVerificationController::class, 'showNotice'])->name('verification.notice');
         Route::post('resend', [EmailVerificationController::class, 'resend'])->middleware(['throttle:6,1'])->name('verification.send');
@@ -15,17 +21,17 @@ Route::prefix('auth')->group(function () {
     });
 });
 
-// Login and dashboard routes
+// Login and register routes
 Route::middleware('guest')->group(function () {
 
     // Authentication routes
     Route::prefix('auth')->group(function () {
 
-        Route::post('login', [LoginController::class, 'authenticate'])->middleware('guest');
-        Route::post('register', [RegisterController::class, 'authenticate'])->middleware('guest');
+        Route::post('login', [LoginController::class, 'authenticate']);
+        Route::post('register', [RegisterController::class, 'authenticate']);
     });
 
     // Index routes for views
-    Route::get('login', [LoginController::class, 'index'])->middleware('guest')->name('login');
-    Route::get('register', [RegisterController::class, 'index'])->middleware('guest')->name('register');
+    Route::get('login', [LoginController::class, 'index'])->name('login');
+    Route::get('register', [RegisterController::class, 'index'])->name('register');
 });